Improved Action support
Modules can now perform ACTION and SAY within the same code block. From within the exports.execute function of a module, pass a boolean value as a second parameter to the callback function. This will be interpreted as isAction = true in the bot's execute function, and encode the message correctly before passing it back to the JerkBot/IRC-js say function.
See slap.js for example usage.

Also added a makeAction util function to create ACTIONs out of messages.